1. Avoid removing soft link of an empty directory 2. Print more info when 'make install'/'make uninstall' 3. Add comments to the templates

+++ build/sh/cleandir.sh
@@ -1,5 +1,16 @@
 #!/bin/bash
 for dir in $@ ; do
+	# Don't delete soft link. Althrogh the commented code can safely delete
+	# the soft link of an empty directory. But I still think recreating a
+        # soft link is far too 'difficult' than recreating a soft link. So, keep
+	# the soft link.
+	if [ -h $dir ]; then
+		continue
+		#if [ ! "`ls -A $dir`" ]; then
+		#	rm $dir
+		#fi
+	fi
+
 	if [ -d $dir ]; then
 		rmdir -p --ignore-fail-on-non-empty $dir
 	fi
